Add files via upload
The script imports Google stocks historical data and converts it's elements from string to ints and floats as needed and then appends them to an empty list. This empty list creates a list of lists, where each list is the data for one day. Then it calculates the daily return value for every day and prints it. The script accomplishes the later by creating values by indexing the list at various parts and then executing the values by using a given formula.